+++ src/actor/actor_data.cpp
@@ -1,0 +1,5 @@
+﻿#include "actor_data.h"
+
+namespace actor {
+    
+}  // actor
