--- src/event.h
@@ -1,9 +1,0 @@
-#pragma once
-#include "event/keymanager.h"
-
-namespace event {
-    inline void sink_event_handlers() {
-        key_manager::sink();
-        logger::info("Start listening for InputEvents."sv);
-    }
-}
