+++ src/event/event.cpp
@@ -1,0 +1,9 @@
+﻿#include "event.h"
+#include "event/input_event.h"
+
+namespace event {
+    void event::sink_event_handler() {
+        input_event::sink();
+        logger::info("added input event sink"sv);
+    }
+}  // event
