+++ src/event/event.h
@@ -1,0 +1,8 @@
+﻿#pragma once
+
+namespace event {
+    class event {
+    public:
+        static void sink_event_handler();
+    };
+}  // event
