+++ src/hook/hook.cpp
@@ -1,0 +1,8 @@
+﻿#include "hook.h"
+#include "hook/dialogue_menu_hook.h"
+
+namespace hook {
+    void hook::install() {
+        dialogue_menu_hook::install();
+    }
+}  // hook
