+++ src/hook/hook.h
@@ -1,0 +1,9 @@
+﻿#pragma once
+
+namespace hook {
+    class hook {
+    public:
+        static void install();
+    };
+}  // hook
+
