+++ src/scaleform/scaleform.cpp
@@ -1,0 +1,9 @@
+﻿#include "scaleform.h"
+#include "scaleform/menu/dialogue_info_menu.h"
+
+namespace scaleform {
+    void scaleform::Register() {
+        dialogue_info_menu::Register();
+        logger::info("registered all forms");
+    }
+}  // scaleform
