+++ src/util/offset.h
@@ -2,5 +2,4 @@
 //ae might need fixing
 namespace offset {
     constexpr REL::RelocationID get_relationship_rank(53898, 54710);
-    constexpr REL::RelocationID get_dialog_menu(268589, 215255);
 }
